requirements.txt
amqp==2.4.0
backcall==0.1.0
beautifulsoup4==4.6.3
billiard==3.5.0.5
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
Django==2.1.2
django-appconf==1.0.2
django-debug-toolbar==1.11
django-jet==1.0.8
django-postgres-copy==2.3.7
django-select2==6.3.1
django-session-timeout==0.0.4
et-xmlfile==1.0.1
idna==2.7
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.13.1
kombu==4.2.2.post1
numpy==1.15.4
openpyxl==2.5.9
parso==0.3.1
pdftotext==2.1.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
redis==2.10.6
requests==2.20.1
six==1.12.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
xlrd==1.2.0
xlwt==1.3.0
